// src/domain/enums/badge-type.enum.ts
/**
 * 뱃지 타입 열거형
 * 뱃지 획득 조건의 타입을 정의
 */
export enum BadgeType {
  FIRST_CHALLENGE = 'FIRST_CHALLENGE',
  STREAK_7 = 'STREAK_7',
  STREAK_30 = 'STREAK_30',
  PERFECT_WEEK = 'PERFECT_WEEK',
  PERFECT_MONTH = 'PERFECT_MONTH',
  EARLY_BIRD = 'EARLY_BIRD',
  NIGHT_OWL = 'NIGHT_OWL',
  SOCIAL_BUTTERFLY = 'SOCIAL_BUTTERFLY',
  ENCOURAGER = 'ENCOURAGER',
  CHALLENGE_MASTER = 'CHALLENGE_MASTER',
} 